fix: require cgo for Open
We still allow users to reference the rest of the types without CGO to
make things easier for users. But we put Open (and the tests) behind a
cgo build flag.

This will let us deploy our unified CI.
